Added a method to clock for a max amount of cycles
