+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "caklimas-rust-gameboy"
-version = "0.4.3"
+version = "0.4.4"
 authors = ["caklimas <caklimas@gmail.com>"]
 edition = "2021"
