+++ src/files/package.json
@@ -3,7 +3,7 @@
   "collaborators": [
     "caklimas <caklimas@gmail.com>"
   ],
-  "version": "0.4.3",
+  "version": "4",
   "files": [
     "caklimas_rust_gameboy_bg.wasm",
     "caklimas_rust_gameboy_bg.js",
