+++ src/lib.rs
@@ -1,10 +1,7 @@
-use std::{cell::RefCell, rc::Rc};
-
 use gameboy::Gameboy;
 
-use wasm_bindgen::{prelude::*, JsCast};
-use web::Emulator;
-use web_sys::{AudioContext, Window};
+use input::Input;
+use wasm_bindgen::prelude::*;
 
 #[macro_use]
 extern crate bitfield;
@@ -24,22 +21,57 @@ pub mod input;
 pub mod mbc;
 pub mod mmu;
 pub mod utils;
-pub mod web;
 
 #[cfg(feature = "wee_alloc")]
 #[global_allocator]
 static ALLOC: wee_alloc::WeeAlloc = wee_alloc::WeeAlloc::INIT;
 
 #[wasm_bindgen]
-extern "C" {
-    // Use `js_namespace` here to bind `console.log(..)` instead of just
-    // `log(..)`
-    #[wasm_bindgen(js_namespace = console)]
-    fn log(s: &str);
+pub struct Emulator {
+    cycles: usize,
+    gameboy: Gameboy,
 }
 
-macro_rules! console_log {
-    // Note that this is using the `log` function imported above during
-    // `bare_bones`
-    ($($t:tt)*) => (log(&format_args!($($t)*).to_string()))
+#[wasm_bindgen]
+impl Emulator {
+    #[wasm_bindgen(constructor)]
+    pub fn new(bytes: Vec<u8>) -> Self {
+        let gameboy = Gameboy::new(bytes, true);
+        Self { cycles: 0, gameboy }
+    }
+
+    pub fn clock_until_event(&mut self, max_cycles: usize) -> EmulatorState {
+        while self.cycles < max_cycles {
+            let (cycles, audio_full) = self.gameboy.clock();
+            self.cycles += cycles as usize;
+            if audio_full {
+                return EmulatorState::AudioFull;
+            } else if self.gameboy.frame_complete() {
+                return EmulatorState::FrameComplete;
+            }
+        }
+
+        self.cycles -= max_cycles;
+        EmulatorState::MaxCycles
+    }
+
+    pub fn update_controls(&mut self, input: Input) {
+        self.gameboy.update_controls(input);
+    }
+
+    pub fn get_screen(&self) -> Vec<u8> {
+        self.gameboy.get_screen().to_vec()
+    }
+
+    pub fn get_audio_buffer(&self) -> Vec<f32> {
+        self.gameboy.get_audio_buffer().to_vec()
+    }
+}
+
+#[wasm_bindgen]
+pub enum EmulatorState {
+    None,
+    FrameComplete,
+    AudioFull,
+    MaxCycles,
 }
